Fornorsket alt i frontend-kodebasen
