+++ frontend/src/containere/readme.md
@@ -1,0 +1,9 @@
+# Containers
+
+This folder should contain all LOGICAL/STATE components
+
+-   Can have app state
+-   Can perform remote actions
+-   Can be a page component
+-   Can have presentation
+-   Can include view/presentation components
