+++ frontend/src/komponenter/hovedomrade/Hovedomrade.tsx
