+++ frontend/src/komponenter/readme.md
@@ -1,0 +1,17 @@
+# Components
+
+This folder should contain all VIEW/PRESENTATIONAL components which are used
+through out the application
+
+-   Can have props
+-   Can have local state
+-   Can NOT have app state
+-   Can NOT perform remote actions# Components
+
+This folder should contain all VIEW/PRESENTATIONAL components which are used
+through out the application
+
+-   Can have props
+-   Can have local state
+-   Can NOT have app state
+-   Can NOT perform remote actions
