--- frontend/src/style/less/components/hovedomrade/hovedinnhold.less
@@ -1,20 +1,0 @@
-.hovedinnhold {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 2.5rem;
-    height: 29.6rem;
-}
-
-.logininfo {
-    margin-top: 15% !important;
-    margin-bottom: 8% !important;
-    white-space: nowrap;
-}
-
-.fremgangsknapper {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    margin-top: auto;
-}
