--- frontend/src/style/less/components/navigation/footer.less
@@ -1,34 +1,0 @@
-@import (reference) '~nav-frontend-core/less/_variabler';
-
-.footer {
-    background: @navGra20;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    position: fixed;
-    bottom: 0;
-}
-
-.logo_info {
-    display: flex;
-    align-items: center;
-    margin-left: 1.4375rem;
-}
-
-.logo {
-    margin: 1.125rem;
-    height: 2.5rem;
-    width: 4rem;
-}
-
-.lenkeBottom {
-    color: @navMorkGra !important;
-    margin-right: 3rem;
-}
-
-.lenkeBottom:active,
-.lenkeBottom:focus {
-    color: white !important;
-    background-color: black !important;
-}
