--- frontend/src/style/less/components/navigation/topBar.less
@@ -1,18 +1,0 @@
-@import (reference) '~nav-frontend-core/less/_variabler';
-
-.topBar {
-    background: @navOransjeLighten60;
-    display: flex;
-    position: fixed;
-    justify-content: space-between;
-    align-items: center;
-    height: 88px;
-    width: 100%;
-    padding: 0 2rem;
-}
-
-.sidetittel {
-    position: absolute;
-    left: 50%;
-    transform: translateX(-50%);
-}
