+++ frontend/src/style/less/containere/brukerKnappToppBar.less
@@ -1,0 +1,24 @@
+@import (reference) '~nav-frontend-core/less/_variabler';
+
+.personIkon {
+    width: 1.375rem !important;
+    height: 1.5rem !important;
+    margin-left: 0.75rem !important;
+}
+
+.brukerKnapp {
+    color: black !important;
+    margin-left: 2rem;
+}
+
+.brukerKnapp:hover {
+    color: #0067c5 !important;
+}
+
+.brukerKnapp:focus {
+    color: #0067c5 !important;
+}
+
+.brukerKnapp:active {
+    color: #fff !important;
+}
