+++ frontend/src/style/less/containere/loggUtKnappToppBar.less
@@ -1,0 +1,16 @@
+.loggUtKnapp {
+    color: black !important;
+    margin-right: 2rem;
+}
+
+.loggUtKnapp:hover {
+    color: #0067c5 !important;
+}
+
+.loggUtKnapp:focus {
+    color: #0067c5 !important;
+}
+
+.loggUtKnapp:active {
+    color: #fff !important;
+}
