--- frontend/src/style/less/containers/loginButtonTopBar.less
@@ -1,24 +1,0 @@
-@import (reference) '~nav-frontend-core/less/_variabler';
-
-.personIkon {
-    width: 1.375rem !important;
-    height: 1.5rem !important;
-    margin-left: 0.75rem !important;
-}
-
-.loginBtn1 {
-    color: black !important;
-    margin-left: 2rem;
-}
-
-.loginBtn1:hover {
-    color: #0067c5 !important;
-}
-
-.loginBtn1:focus {
-    color: #0067c5 !important;
-}
-
-.loginBtn1:active {
-    color: #fff !important;
-}
