--- frontend/src/style/less/containers/logoutButtonTopBar.less
@@ -1,16 +1,0 @@
-.logoutBtn1 {
-    color: black !important;
-    margin-right: 2rem;
-}
-
-.logoutBtn1:hover {
-    color: #0067c5 !important;
-}
-
-.logoutBtn1:focus {
-    color: #0067c5 !important;
-}
-
-.logoutBtn1:active {
-    color: #fff !important;
-}
