+++ frontend/src/style/less/komponenter/hovedomrade/hovedomrade.less
