+++ frontend/src/style/less/komponenter/hovedomrade/hovedomradeTopp.less
@@ -1,0 +1,25 @@
+@import (reference) '~nav-frontend-core/less/_variabler';
+
+.hovedomradeTopp {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+}
+
+.toppIkon {
+    height: 2.25rem;
+    width: 2.25rem;
+    position: relative;
+    top: -1rem;
+}
+
+.infotekst {
+    margin: 1rem 2.5rem 0.9375rem 2.5rem !important;
+    white-space: nowrap;
+}
+
+.divider {
+    background: @navOransjeLighten60;
+    height: 4px;
+    width: 100%;
+}
