+++ frontend/src/style/less/komponenter/navigasjon/footer.less
@@ -1,0 +1,34 @@
+@import (reference) '~nav-frontend-core/less/_variabler';
+
+.footer {
+    background: @navGra20;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    position: fixed;
+    bottom: 0;
+}
+
+.logoInfo {
+    display: flex;
+    align-items: center;
+    margin-left: 1.4375rem;
+}
+
+.logo {
+    margin: 1.125rem;
+    height: 2.5rem;
+    width: 4rem;
+}
+
+.lenkeBunn {
+    color: @navMorkGra !important;
+    margin-right: 3rem;
+}
+
+.lenkeBunn:active,
+.lenkeBunn:focus {
+    color: white !important;
+    background-color: black !important;
+}
