+++ frontend/src/style/less/komponenter/navigasjon/toppBar.less
@@ -1,0 +1,18 @@
+@import (reference) '~nav-frontend-core/less/_variabler';
+
+.toppBar {
+    background: @navOransjeLighten60;
+    display: flex;
+    position: fixed;
+    justify-content: space-between;
+    align-items: center;
+    height: 88px;
+    width: 100%;
+    padding: 0 2rem;
+}
+
+.sidetittel {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+}
