+++ frontend/src/style/less/main.less
@@ -2,21 +2,21 @@
 @import (reference) '~nav-frontend-typografi-style/src/index.less';
 @import 'nav-frontend-core';
 
-//COMPONENTS
-@import './components/hovedomrade/hovedinnhold.less';
-@import './components/hovedomrade/hovedomrade.less';
-@import './components/hovedomrade/hovedomradeTopp.less';
-@import './components/navigation/footer.less';
-@import './components/navigation/topBar.less';
+//KOMPONENTER
+@import './komponenter/hovedomrade/hovedinnhold.less';
+@import './komponenter/hovedomrade/hovedomrade.less';
+@import './komponenter/hovedomrade/hovedomradeTopp.less';
+@import './komponenter/navigasjon/footer.less';
+@import './komponenter/navigasjon/toppBar.less';
 
-//CONTAINERS
-@import './containers/loginButtonTopBar.less';
-@import './containers/loginPrimer.less';
-@import './containers/logoutButtonTopBar.less';
+//CONTAINERE
+@import './containere/brukerKnappToppBar.less';
+@import './containere/loggInnHovedknapp.less';
+@import './containere/loggUtKnappToppBar.less';
 
-//VIEWS
-@import './views/landingsside.less';
-@import './views/samtykkeskjema.less';
+//VISNINGER
+@import './visninger/landingsside.less';
+@import './visninger/samtykkeskjema.less';
 
 body > #root > div {
     height: 100vh;
