--- frontend/src/views/landingsside/Landingsside.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import Hovedomrade from '../../components/hovedomrade/Hovedomrade'
-
-export default function Landingsside(): React.ReactElement {
-    return (
-        <div className="landingsside">
-            <Hovedomrade visFremgangsknapper={false} tittel={'Velkommen til NAVs innbyggerpanel'} />
-        </div>
-    )
-}
