+++ frontend/src/visninger/landingsside/Landingsside.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Hovedomrade from '../../komponenter/hovedomrade/Hovedomrade'
+
+export default function Landingsside(): React.ReactElement {
+    return (
+        <div className="landingsside">
+            <Hovedomrade visFremgangsknapper={false} tittel={'Velkommen til NAVs innbyggerpanel'} />
+        </div>
+    )
+}
